Working and writing to usage db. Ready to do a cleanup

+++ ui/dfuicommthread.cpp
@@ -89,7 +89,7 @@ QByteArray DfUiCommThread::readyRead()
     {
         QString temperature = Data.mid(13);
         qDebug() << temperature;
-        emit temperatureSignal(temperature);
+        emit temperatureSignal(temperature.toDouble());
     }
     // else if (strtol(Data, &pEnd, 10) || (Data[0] == '0' && Data[1] == '.'))
     // old rudimentary way of sending dispensed volumen from fsm to ui
